split models to entities and services
